TableDiffFlask/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "datacompy>=0.18.1",
    "pandas>=2.3.3",
    "pymssql>=2.3.0",
    "python-docx>=1.2.0",
    "snowflake-connector-python>=4.0.0",
]
